refactor(wallets): declare each wallet type once
`InstallObjects` and `NamespaceData` existed twice, so `@rango-dev/ui` and the
widget's stateful connect read them from wallets-react, which declares them.
`Persistor` regains its `PersistStorage` interface and moves beside its caller
under `hub`.

The demo imports `WalletType` from hub3js directly and stops passing arguments
to `allProviders`, which takes none. `rango-preset` lists wallets-react as a
devDependency alongside the peer one.
