+++ queue-manager/queue-manager-demo/package.json
@@ -19,6 +19,7 @@
     "rango-sdk": "^0.5.0"
   },
   "dependencies": {
+    "@hub3js/core": "0.0.0-experimental-508611c9-20260825",
     "@rango-dev/internal-blockchains": "^0.1.1-next.0",
     "@rango-dev/provider-all": "^0.65.1-next.3",
     "@rango-dev/queue-manager-rango-preset": "^0.64.1-next.3",
