+++ queue-manager/queue-manager-demo/src/components/FlowsList/index.tsx
@@ -1,4 +1,4 @@
-import type { WalletType } from '@rango-dev/queue-manager-rango-preset';
+import type { WalletType } from '@hub3js/core';
 
 import { useManager } from '@rango-dev/queue-manager-react';
 import React, { useEffect, useState } from 'react';
