+++ wallets/react/src/hub/mod.ts
@@ -1,2 +1,2 @@
-export { getHubProviders, findProviderByType } from './utils.js';
+export { getHubProviders, findProviderByType, HUB_VERSION } from './utils.js';
 export { useHubAdapter } from './useHubAdapter.js';
