+++ wallets/react/src/index.ts
@@ -1,3 +1,4 @@
+export { HUB_VERSION } from './hub/mod.js';
 export { default as Provider } from './provider.js';
 export { useWallets } from './legacy/hooks.js';
 export * from './legacy/types.js';
