+++ widget/playground/src/containers/FunctionalLayout/FunctionalLayout.utils.ts
@@ -5,11 +5,10 @@ import type { BlockchainMeta } from 'rango-sdk';
 import { pickVersion, type VersionedProviders } from '@hub3js/core/utils';
 import { getSupportedChainsFromProvider } from '@rango-dev/internal-blockchains';
 import { allProviders as getAllProviders } from '@rango-dev/provider-all';
+import { HUB_VERSION } from '@rango-dev/wallets-react';
 
 import { getCategoryNetworks } from '../../utils/blockchains';
 import { excludedWallets } from '../../utils/common';
-
-const HUB_VERSION = '1.0.0';
 
 // Considering that the wallets list of the config used for `WidgetProvider` gets filtered by the selected wallets here, we can not directly use `getWalletInfo` of `useWallets` to get the info related to each wallet item because the required provider for an unselected wallet item will not be passed to `Provider`.
 export function getWalletsList(
