+++ widget/ui/package.json
@@ -46,6 +46,7 @@
     "react-is": "^18.2.0"
   },
   "dependencies": {
+    "@rango-dev/wallets-react": "^0.48.1-next.3",
     "@hub3js/core": "0.0.0-experimental-508611c9-20260825",
     "@radix-ui/react-checkbox": "^1.0.1",
     "@radix-ui/react-collapsible": "^1.0.3",
