+++ widget/ui/src/types/index.ts
@@ -1,10 +1,4 @@
-export type {
-  ConnectedWallet,
-  InstallObjects,
-  NamespaceMeta,
-  NeedsDerivationPath,
-  NeedsNamespace,
-} from './wallet.js';
+export type { ConnectedWallet } from './wallet.js';
 export type {
   SimulationAssetAndAmount,
   SimulationValidationStatus,
