+++ widget/ui/src/types/wallet.ts
@@ -1,20 +1,4 @@
-import type {
-  DerivationPathProperty,
-  NamespacesProperty,
-  WalletType,
-} from '@hub3js/core';
-
-export type NamespaceMeta = NamespacesProperty['value']['data'][number];
-export type NeedsNamespace = NamespacesProperty['value'];
-export type NeedsDerivationPath = DerivationPathProperty['value'];
-
-export type InstallObjects = {
-  CHROME?: string;
-  FIREFOX?: string;
-  EDGE?: string;
-  BRAVE?: string;
-  DEFAULT: string;
-};
+import type { WalletType } from '@hub3js/core';
 
 interface Wallet {
   chain: string;
